Some changes

+++ lib/segmentation.h
@@ -40,8 +40,8 @@ uint32 seg_init_regs(uint8 *r, uint8 *g, uint8 *b, uint8 *r1, uint16 *rg, uint32
 uint32 seg_fill_regs(uint8 *r, uint8 *g, uint8 *b, uint8 *r1, uint16 *rg, uint32 *col, uint32 *l1, uint32 rgc, uint32 w, uint32 h);
 
 void seg_fall_forest(uint8 *img, uint8 *img1, uint32 w, uint32 h);
-uint32 seg_group_pixels(uint8 *r, uint8 *g, uint8 *b, uint8 *r1, uint16 *rg, uint8 *col, uint32 *l1, uint32 *l2, uint32 w, uint32 h);
-void seg_draw_reg(uint8 *r, uint8 *g, uint8 *b, uint16 *rg, uint8 *col, uint32 w, uint32 h);
+uint32 seg_group_pixels(uint8 *r, uint8 *g, uint8 *b, uint8 *r1, uint32 *rg, uint8 *col, uint32 *l1, uint32 *l2, uint32 w, uint32 h);
+void seg_draw_reg(uint8 *r, uint8 *g, uint8 *b, uint32 *rg, uint8 *col, uint32 w, uint32 h);
 
 void seg_find_clusters(uint32 *i3d, uint16 *lut, int16 *r, int16 *g, int16 *b, uint32 w, uint32 h, uint32 rd,  uint32 bpp, p3d *q,uint32 *buf);
 void seg_quantization(uint16 *lut, uint8 *rgb, int16 *r, int16 *g, int16 *b, uint32 w, uint32 h, uint32 bpp, p3d *q);
